+++ src/jinja2/runtime.py
@@ -5,9 +5,7 @@ import typing as t
 from collections import abc
 from itertools import chain
 
-from markupsafe import escape  # noqa: F401
-from markupsafe import Markup
-from markupsafe import soft_str
+import markupsafe
 
 from .async_utils import auto_aiter
 from .async_utils import auto_await  # noqa: F401
@@ -17,6 +15,7 @@ from .exceptions import UndefinedError
 from .nodes import EvalContext
 from .utils import _PassArg
 from .utils import concat
+from .utils import EscapeFunc
 from .utils import internalcode
 from .utils import missing
 from .utils import Namespace  # noqa: F401
@@ -41,15 +40,17 @@ if t.TYPE_CHECKING:
             ...
 
 
+html_escape = markupsafe.escape
+
 # these variables are exported to the template runtime
 exported = [
     "LoopContext",
     "TemplateReference",
     "Macro",
-    "Markup",
     "TemplateRuntimeError",
     "missing",
-    "escape",
+    "concat",
+    "html_escape",
     "markup_join",
     "str_join",
     "identity",
@@ -72,19 +73,33 @@ def identity(x: V) -> V:
     return x
 
 
-def markup_join(seq: t.Iterable[t.Any]) -> str:
-    """Concatenation that escapes if necessary and converts to string."""
+def markup_join(
+    seq: t.Iterable[t.Any], mark_safe: EscapeFunc = markupsafe.Markup
+) -> t.Union[str, markupsafe.Markup]:
+    """
+    Concatenation that escapes if necessary and converts to string.
+
+    .. versionchanged:: 3.1
+        added optional parameter escape_function to make
+        use the context based escape function
+    """
     buf = []
-    iterator = map(soft_str, seq)
+    iterator = map(markupsafe.soft_str, seq)
     for arg in iterator:
         buf.append(arg)
         if hasattr(arg, "__html__"):
-            return Markup("").join(chain(buf, iterator))
+            return mark_safe("").join(chain(buf, iterator))
     return concat(buf)
 
 
-def str_join(seq: t.Iterable[t.Any]) -> str:
-    """Simple args to string conversion and concatenation."""
+def str_join(seq: t.Iterable[t.Any], mark_safe: EscapeFunc = markupsafe.Markup) -> str:
+    """
+    Simple args to string conversion and concatenation.
+
+    .. versionchanged:: 3.1
+            added optional and currently ignored parameter
+            ``mark_safe`` to allow easier usage of ``markup_join``
+    """
     return concat(map(str, seq))
 
 
@@ -367,7 +382,7 @@ class BlockReference:
         )
 
         if self._context.eval_ctx.autoescape:
-            return Markup(rv)
+            return self._context.eval_ctx.mark_safe(rv)
 
         return rv
 
@@ -379,8 +394,7 @@ class BlockReference:
         rv = concat(self._stack[self._depth](self._context))
 
         if self._context.eval_ctx.autoescape:
-            return Markup(rv)
-
+            rv = self._context.eval_ctx.mark_safe(rv)
         return rv
 
 
@@ -667,6 +681,7 @@ class Macro:
         default_autoescape: t.Optional[bool] = None,
     ):
         self._environment = environment
+        self._mark_safe: EscapeFunc = environment.get_markup_class()
         self._func = func
         self._argument_count = len(arguments)
         self.name = name
@@ -703,8 +718,14 @@ class Macro:
         # argument to callables otherwise anyway.  Worst case here is
         # that if no eval context is passed we fall back to the compile
         # time autoescape flag.
+
         if args and isinstance(args[0], EvalContext):
             autoescape = args[0].autoescape
+            # If the eval context is available we use it to determine
+            # the correct mark safe method
+            # otherwise mark safe is already set in the __init__
+            # function from environmental context
+            self._mark_safe = args[0].mark_safe
             args = args[1:]
         else:
             autoescape = self._default_autoescape
@@ -766,7 +787,7 @@ class Macro:
         rv = await self._func(*arguments)  # type: ignore
 
         if autoescape:
-            return Markup(rv)
+            return self._mark_safe(rv)
 
         return rv  # type: ignore
 
@@ -777,8 +798,7 @@ class Macro:
         rv = self._func(*arguments)
 
         if autoescape:
-            rv = Markup(rv)
-
+            rv = self._mark_safe(rv)
         return rv
 
     def __repr__(self) -> str: