+++ tests/conftest.py
@@ -47,3 +47,29 @@ def choice_loader(dict_loader, package_loader):
 def prefix_loader(filesystem_loader, dict_loader):
     """returns a PrefixLoader"""
     return loaders.PrefixLoader({"a": filesystem_loader, "b": dict_loader})
+
+
+@pytest.fixture
+def custom_escape_func():
+    """Return a simple custom escape function"""
+
+    def dollar_to_eur(s):
+        return str(s).replace("$", "€")
+
+    return dollar_to_eur
+
+
+@pytest.fixture
+def return_custom_autoescape(custom_escape_func):
+    """return a simple example for a custom escape function"""
+
+    def do_return_autoescape(suffix):
+        return custom_escape_func
+
+    return do_return_autoescape
+
+
+@pytest.fixture
+def env_custom_autoescape(return_custom_autoescape):
+    """return a simple example for a custom escape function"""
+    return Environment(autoescape=return_custom_autoescape)