+++ tests/test_filters.py
@@ -179,6 +179,17 @@ class TestFilter:
         t = env.from_string('{{ "jinja"|indent(blank=true) }}')
         assert t.render() == "jinja"
 
+    def test_indent_custom_escape(self, env_custom_autoescape):
+        env = env_custom_autoescape
+        text = "\n".join(["", "foo$bar", '"baz"', ""])
+        ident_default = "{{ foo|indent(2, false, false) }}"
+        t = env.from_string(ident_default)
+        assert t.render(foo=text) == '\n  foo€bar\n  "baz"\n'
+        text_secure = Markup(text)
+        assert t.render(foo=text_secure) == '\n  foo$bar\n  "baz"\n'
+        text_secure = env.default_markup_class(text)
+        assert t.render(foo=text_secure) == '\n  foo$bar\n  "baz"\n'
+
     def test_indent_markup_input(self, env):
         """
         Tests cases where the filter input is a Markup type
@@ -227,9 +238,15 @@ class TestFilter:
         out = tmpl.render()
         assert out == "1|2|3"
 
+    def test_join_default_autoescape(self):
         env2 = Environment(autoescape=True)
         tmpl = env2.from_string('{{ ["<foo>", "<span>foo</span>"|safe]|join }}')
         assert tmpl.render() == "&lt;foo&gt;<span>foo</span>"
+
+    def test_join_custom_autoescape(self, return_custom_autoescape):
+        env2 = Environment(autoescape=return_custom_autoescape)
+        tmpl = env2.from_string('{{ ["$foo$", "$span$foo$span$"|safe]|join }}')
+        assert tmpl.render() == "€foo€$span$foo$span$"
 
     def test_join_attribute(self, env):
         User = namedtuple("User", "username")
@@ -656,9 +673,28 @@ class TestFilter:
         tmpl = env.from_string('{{ string|replace("o", ">x<") }}')
         assert tmpl.render(string=Markup("foo")) == "f&gt;x&lt;&gt;x&lt;"
 
+    def test_replace_custom_escape(self, env_custom_autoescape):
+        env = Environment()
+        tmpl = env.from_string('{{ string|replace("o", 42) }}')
+        assert tmpl.render(string="<$foo>$") == "<$f4242>$"
+        env = env_custom_autoescape
+        tmpl = env.from_string('{{ string|replace("o", 42) }}')
+        assert tmpl.render(string="<$foo>$") == "<€f4242>€"
+        tmpl = env.from_string('{{ string|replace("<$", 42) }}')
+        assert tmpl.render(string="<$foo>$") == "42foo>€"
+        tmpl = env.from_string('{{ string|replace("o", ">$x<$") }}')
+        assert tmpl.render(string=env.default_markup_class("foo")) == "f>€x<€>€x<€"
+        tmpl = env.from_string('{{ string|replace("o", ">$x<$") }}')
+        assert tmpl.render(string=Markup("foo")) == "f>€x<€>€x<€"
+
     def test_forceescape(self, env):
         tmpl = env.from_string("{{ x|forceescape }}")
         assert tmpl.render(x=Markup("<div />")) == "&lt;div /&gt;"
+
+    def test_foceescape_custom(self, env_custom_autoescape):
+        env = env_custom_autoescape
+        tmpl = env.from_string("{{ x|forceescape }}")
+        assert tmpl.render(x=Markup("<div$ />")) == "<div€ />"
 
     def test_safe(self, env):
         env = Environment(autoescape=True)
@@ -666,6 +702,13 @@ class TestFilter:
         assert tmpl.render() == "<div>foo</div>"
         tmpl = env.from_string('{{ "<div>foo</div>" }}')
         assert tmpl.render() == "&lt;div&gt;foo&lt;/div&gt;"
+
+    def test_safe_custom_escape(self, env_custom_autoescape):
+        env = env_custom_autoescape
+        tmpl = env.from_string('{{ "<div>$foo$</div>"|safe }}')
+        assert tmpl.render() == "<div>$foo$</div>"
+        tmpl = env.from_string('{{ "<div>$foo$</div>" }}')
+        assert tmpl.render() == "<div>€foo€</div>"
 
     @pytest.mark.parametrize(
         ("value", "expect"),